[#25] Use a padding for the camera position against walls/floor
I found out how to properly use the collision normal - it just has to be transformed properly by the object's transformation matrix. However, I don't use it; I only decrease the camera distance instead because 1) it's simpler to implement, 2) it doesn't lead to strange jittery movements of the camera, and 3) it doesn't change the angle in relation to the camera's target.