Improve the map rendering
Optimize how pixel colour is determined
1. Loop only through polygons which are relevant to the given cell (~8x improvement for a 30x30 map, but infinite for an infinitely big map)
2. Wrap the point around only if it's at the edge of the map (~9x improvement)

Fix the fertility calculation - instead of `1 - (distance % 2) / (requiredDistance ^ 2)`, it should be `(1 - distance / requiredDistance) ^ 2`.

Improve the quality of the image by drawing the lines of the polygons and the centers of the cells only after the pixels have been filled so that antialiasing can be used.