[Close #62] Have only the necessary forest cells in the scene
Use a pool of forest cell objects that is dynamically enlarged when necessary. With the current 50x50 map, there are less than 70 cells at once (way less than the previous 2500!).

There is roughly a ~10ms frame delay improvement in the current world.

In a world that is 9x bigger (150x150 cells), there isn't a difference - both are abysmal (~3 FPS) despite how this change drastically reduces the number of cells in the scene. Maybe there's another factor at play here. The new version is still better though.

By the way, a 250x250 world cannot be generated because it consumes too much memory. That is 62500 total cells. I thought them reusing data would limit the number to around ~250 unique cells regardless of the total number of cells, but it seems not to be enough. Even with the new code, the consumed memory is too much.