[#67] Create the 'core' and 'editor' modules
The 'core' module contains everything that was in the 'backend' module which isn't related to the server but rather to the game.

The 'editor' module contains the world editor code. The backend does not depend on it so making changes to it should not cause the whole project to have to be rebuilt.