[#41] Optimize the server preview
Now, it doesn't redraw anything if nothing has changed. As a nice side effect, I also simplified the code in the process, which allowed me to remove the suppression of a complexity warning.

Also, I made it so that the player name and the player latency are on two separate lines and have a semi-transparent background.